修正清除通知的确认对话框无法取消的 Bug;

+++ www/js/controllers/notifications.controller.js
@@ -68,9 +68,11 @@
 
     function deleteAllNotifications() {
       BaseService.confirm("清除通知", "", "确定要将你的所有通知清空吗？").then(function(res) {
-        UserService.deleteAllNotifications().then(function(res) {
-          doRefresh();
-        });
+        if (res) {
+          UserService.deleteAllNotifications().then(function(res) {
+            doRefresh();
+          });
+        }
       });
     }
